+++ farms-liquibase/scripts/01_00_xx/01_00_02/06/ddl/routines/farms_import_pkg.append_imp1.sql
@@ -1,11 +1,11 @@
 create or replace procedure farms_import_pkg.append_imp1(
-   in in_version_id numeric,
+   in in_version_id bigint,
    in in_msg varchar
 )
 language plpgsql
 as $$
 declare
-    eatit numeric;
+    eatit bigint;
 begin
     eatit := farms_import_pkg.append_imp(in_version_id, in_msg);
 end;
